# portlist: update some internals to use append-style APIs